♻️ update serve view
