+++ src/serves/backend/ceobe_operation/resource/mod.rs
@@ -1,3 +1,3 @@
 mod controller;
 mod error;
-mod view_models;
+mod view;
